+++ R/install.R
@@ -1,8 +1,8 @@
 # --- FILE: localLLM/R/install.R ---
 
 # Define library version and base URL
-.lib_version <- "1.2.0"
-.base_url <- "https://github.com/EddieYang211/localLLM/releases/download/v1.2.0/"
+.lib_version <- "1.2.1"
+.base_url <- "https://github.com/EddieYang211/localLLM/releases/download/v1.2.1/"
 
 # Get path for local library storage
 .lib_path <- function() {
@@ -97,7 +97,7 @@ get_lib_path <- function() {
   filename <- NULL
   if (sys == "Darwin") {
     if (arch == "arm64") filename <- "liblocalllm_macos_arm64.zip"
-    # else if (arch == "x86_64") filename <- "liblocalllm_macos_x64.zip" # Future expansion
+    else if (arch == "x86_64") filename <- "liblocalllm_macos_x64.zip"
   } else if (sys == "Windows") {
     if (arch == "x86-64") filename <- "localllm_windows_x64.zip"
   } else if (sys == "Linux") {